+++ tests/mocks/image_processing_service_mock.go
@@ -1,19 +1,23 @@
 package mocks
 
-import cv "gocv.io/x/gocv"
+import (
+	"fmt"
+	cv "gocv.io/x/gocv"
+	m "tyre-match-backend/db/models"
+)
 
 type ImageProcessingServiceMock struct {
-	ProcessCalls []struct {
-		ID    uint
-		Model string
-	}
-	ProcessError error
+	ProcessTyreImpressionCalls []*m.TyreImpression
+	ProcessTyreImpressionError error
+
+	ProcessTyreModelCalls []*m.TyreModel
+	ProcessTyreModelError error
 
 	SaveStageCalls []struct {
-		ID          uint
-		Model       string
-		FileType    string
-		ResultImage *cv.Mat
+		ID       uint
+		Model    string
+		FileType string
+		Image    *cv.Mat
 	}
 	SaveStageError error
 }
@@ -22,25 +26,35 @@ func NewImageProcessingServiceMock() *ImageProcessingServiceMock {
 	return &ImageProcessingServiceMock{}
 }
 
-func (m *ImageProcessingServiceMock) Process(id uint, model string) error {
-	m.ProcessCalls = append(m.ProcessCalls, struct {
-		ID    uint
-		Model string
-	}{ID: id, Model: model})
-	return m.ProcessError
+func (m *ImageProcessingServiceMock) ProcessTyreImpression(tyreImpression *m.TyreImpression) error {
+	m.ProcessTyreImpressionCalls = append(m.ProcessTyreImpressionCalls, tyreImpression)
+	return m.ProcessTyreImpressionError
 }
 
-func (m *ImageProcessingServiceMock) SaveStage(id uint, model, fileType string, resultImage *cv.Mat) error {
+func (m *ImageProcessingServiceMock) ProcessTyreModel(tyreModel *m.TyreModel) error {
+
+	fmt.Println(">>> TEST")
+	m.ProcessTyreModelCalls = append(m.ProcessTyreModelCalls, tyreModel)
+	return m.ProcessTyreModelError
+}
+
+func (m *ImageProcessingServiceMock) SaveStage(id uint, model, fileType string, image *cv.Mat) error {
 	m.SaveStageCalls = append(m.SaveStageCalls, struct {
-		ID          uint
-		Model       string
-		FileType    string
-		ResultImage *cv.Mat
-	}{ID: id, Model: model, FileType: fileType, ResultImage: resultImage})
+		ID       uint
+		Model    string
+		FileType string
+		Image    *cv.Mat
+	}{ID: id, Model: model, FileType: fileType, Image: image})
 	return m.SaveStageError
 }
 
 func (m *ImageProcessingServiceMock) Reset() {
-	m.ProcessCalls = nil
-	m.ProcessError = nil
+	m.ProcessTyreImpressionCalls = nil
+	m.ProcessTyreImpressionError = nil
+
+	m.ProcessTyreModelCalls = nil
+	m.ProcessTyreModelError = nil
+
+	m.SaveStageCalls = nil
+	m.SaveStageError = nil
 }